monitor/monitor/requirements.txt
RPi.GPIO
cffi
smbus-cffi
